aarch64: port the C++ runtime (libcxxabi + libcxx)
Make the C++ runtime build for aarch64 (it was i386-only).  Both libs are now
arch-aware like libbearssl/libpw — per-arch SIMD flags + musl arch headers keyed
on ${_ARCH}; i386 output unchanged (verified).

- contrib/libcxxabi/Makefile, contrib/libcxx/Makefile: ARCH_SIMD / ARCH_MUSL_INC
  conditionals replace the hardcoded -mno-sse + arch/i386.
- contrib/libcxxabi/cxxabi.cc: operator new/delete took `unsigned` (only correct
  where size_t == unsigned, i.e. i386).  Use size_t — required by the language and
  correct on both 32- and 64-bit (this was the one real porting fix).
- Makefile: un-gate world Step 1a/1b so the C++ runtime builds on aarch64 too.

Verified: build/aarch64/lib/{libcxxabi.a,libcxx.a} build, and a non-trivial C++
program (std::string + std::vector + virtual dispatch + new/delete) compiles and
links to a valid AArch64 ELF against them.  Runtime execution will be exercised
by the objGFX/views desktop (#5), the real consumer.

+++ contrib/libcxxabi/cxxabi.cc
@@ -33,15 +33,17 @@ void __cxa_guard_abort(int *g)   { (void)g; }
 // ---------------------------------------------------------------------------
 // Standard heap allocation operators
 // ---------------------------------------------------------------------------
-void *operator new(unsigned size)              { return malloc(size); }
-void *operator new[](unsigned size)            { return malloc(size); }
+// size_t is the only correct type for operator new (32-bit on i386, 64-bit on
+// aarch64); using `unsigned` only happened to work where size_t == unsigned.
+void *operator new(size_t size)                { return malloc(size); }
+void *operator new[](size_t size)              { return malloc(size); }
 void  operator delete(void *ptr)               { free(ptr); }
 void  operator delete[](void *ptr)             { free(ptr); }
-void  operator delete(void *ptr, unsigned)     { free(ptr); }
-void  operator delete[](void *ptr, unsigned)   { free(ptr); }
+void  operator delete(void *ptr, size_t)       { free(ptr); }
+void  operator delete[](void *ptr, size_t)     { free(ptr); }
 
 // Placement new/delete — no allocation
-void *operator new(unsigned, void *ptr)        { return ptr; }
-void *operator new[](unsigned, void *ptr)      { return ptr; }
+void *operator new(size_t, void *ptr)          { return ptr; }
+void *operator new[](size_t, void *ptr)        { return ptr; }
 void  operator delete(void *, void *)          {}
 void  operator delete[](void *, void *)        {}
